Final Commit
Repaired 2 - Convolution & Blurring
Added 4 - Edge detection

+++ src/core/transformations/correction.h
@@ -12,7 +12,7 @@ public:
     virtual PNM* transform();
 
 protected:
-    inline const int    add(int, double);
+    inline const int    add(double, double);
     inline const double mult(int, double);
     inline const int    power(int, double);
 };
